update player-vehicle assignment

+++ server/execution/web_api/execution.py
@@ -2,7 +2,8 @@ import logging
 
 from flask import Blueprint, Response
 from string_utils import booleanize
-from werkzeug.exceptions import NotFound, BadRequest, InternalServerError
+from werkzeug.exceptions import NotFound, BadRequest, InternalServerError, \
+    Forbidden
 
 import models
 from app_config import csrf, db
@@ -13,6 +14,7 @@ from execution.entities.player import Player
 from execution.services import entityloader
 from execution.services.entityloader import load_location, load_role
 from execution.utils.util import try_get_execution
+from utils.dbo import add_vehicles_to_execution_to_session
 from utils.decorator import required, RequiredValueSource, cache
 
 from sqlalchemy import select
@@ -75,13 +77,34 @@ def create_execution(scenario_id: int, name: str):
         db.session.add(new_execution)
         db.session.commit()
 
-        # Case: first execution of scenario -> reassign wildcard entries in PlayersToVehicle
+        # Case: first execution of scenario -> reassign wildcard entries
         vehicle_list = models.PlayersToVehicleInExecution.query.filter_by(
-            execution_id=0,  # wildcard id
+            execution_id=None,  # wildcard
             scenario_id=scenario_id
         ).all()
-        for vehicle in vehicle_list:
-            vehicle.execution_id = new_execution.id
+
+        if len(vehicle_list) > 0:
+            tan_list = []
+            for vehicle in vehicle_list:
+                vehicle.execution_id = new_execution.id
+                tan_list.append(vehicle.player_tan)
+
+            player_list: list[models.Player] = models.Player.query.filter_by(
+                execution_id=None).all()
+            for player in player_list:
+                # Only assign player which are connected to the scenario
+                if player.tan in tan_list:
+                    player.execution_id = new_execution.id
+        else:
+            template_vehicle = (models.PlayersToVehicleInExecution.query
+                                .filter_by(scenario_id=scenario_id).where(
+                                    models.PlayersToVehicleInExecution.execution_id != None)
+                                .first())
+            assert template_vehicle
+            add_vehicles_to_execution_to_session(new_execution.id, scenario_id,
+                                                 template_vehicle.location_id,
+                                                 template_vehicle.vehicle_name,
+                                                 template_vehicle.travel_time)
 
         db.session.commit()
         logging.info(f"new execution created with id: {new_execution.id}")
@@ -97,14 +120,32 @@ def create_execution(scenario_id: int, name: str):
 @required("execution_id", int, RequiredValueSource.FORM)
 def delete_execution(execution_id: int):
     from execution.run import deactivate_execution
+    # delete player assignment
+    db_p_assignments = (models.PlayersToVehicleInExecution.query
+                        .filter_by(execution_id=execution_id)).all()
+    assert len(db_p_assignments) > 0
+
+    # retrieves all executions stored for a scenario
+    execution_in_scenario_query = select(
+        models.PlayersToVehicleInExecution.execution_id,
+        models.PlayersToVehicleInExecution.scenario_id
+    ).where(models.PlayersToVehicleInExecution.scenario_id == db_p_assignments[0].scenario_id
+    ).distinct(models.PlayersToVehicleInExecution.execution_id)
+
+    execution_list = db.session.execute(execution_in_scenario_query).all()
+
+    for db_p_assignment in db_p_assignments:
+        if len(execution_list) == 1:
+            # keep initial setup when deleting last execution
+            db_p_assignment.execution_id = None
+        else:
+            db.session.delete(db_p_assignment)
+
     # delete Player
     db_players = models.Player.query.filter_by(execution_id=execution_id).all()
-    [db.session.delete(dbo) for dbo in db_players]
-
-    # delete player assignment
-    db_p_assignment = (models.PlayersToVehicleInExecution.query
-                       .filter_by(execution_id=execution_id)).all()
-    [db.session.delete(dbo) for dbo in db_p_assignment]
+    if len(execution_list) > 1:
+        # only delete player if more than one execution exists
+        [db.session.delete(dbo) for dbo in db_players]
 
     # delete execution
     db_execution = models.Execution.query.filter_by(id=execution_id).first()
@@ -202,9 +243,10 @@ def add_new_player(id: int, role: int, vehicle: str):
 @web_api.post("/execution/delete-player")
 @required("id", int, RequiredValueSource.ARGS)
 @required("tan", str, RequiredValueSource.FORM)
-def delete_player(id: int, tan: str):
+@required("vehicle", str, RequiredValueSource.FORM)
+def delete_player(id: int, tan: str, vehicle: str):
     execution = try_get_execution(id)
-    __delete_player(execution, tan)
+    __delete_player(execution, tan, vehicle)
     return Response(status=200)
 
 
@@ -235,61 +277,64 @@ def __get_top_level_locations(execution_id: int):
     return vehicle
 
 
-def __delete_player(execution: Execution, tan: str):
+def __delete_player(execution: Execution, tan: str, vehicle: str):
     from execution.run import remove_player
-    player = models.Player.query.filter_by(tan=tan).first()
-    player_to_vehicle = models.PlayersToVehicleInExecution.query.filter_by(
+    players_to_vehicle = models.PlayersToVehicleInExecution.query.filter_by(
         execution_id=execution.id,
-        player_tan=tan
-    ).first()
+        vehicle_name=vehicle
+    ).all()
 
-    if not player:
-        raise BadRequest(f"Player with tan={tan} not found for "
-                         f"execution={execution.id}")
-    db.session.delete(player)
+    if len(players_to_vehicle) > 1:  # at least one player per vehicle
+        player = models.Player.query.filter_by(tan=tan).first()
+        player_to_vehicle = models.PlayersToVehicleInExecution.query.filter_by(
+            execution_id=execution.id,
+            player_tan=tan
+        ).first()
 
-    if player_to_vehicle:
-        db.session.delete(player_to_vehicle)
+        if not player:
+            raise BadRequest(f"Player with tan={tan} not found for "
+                             f"execution={execution.id}")
+        if player_to_vehicle:
+            db.session.delete(player_to_vehicle)
 
-    db.session.commit()
+        db.session.delete(player)
+        db.session.commit()
 
-    if execution.status is execution.Status.PENDING or execution.Status.RUNNING:
-        execution.players.pop(tan)
-        remove_player(tan)
+        if execution.status is execution.Status.PENDING or execution.Status.RUNNING:
+            execution.players.pop(tan)
+            remove_player(tan)
+    else:
+        raise Forbidden("Unable to delete player. One player per vehicle is required")
 
 
 def __add_new_player_to_execution(execution: Execution, role: int,
                                   vehicle: str):
     from execution.run import register_player  # circular import prevention
 
-    tan = str(unique())
-    # take empty seat of vehicle
-    player_to_vehicle = models.PlayersToVehicleInExecution.query.filter_by(
+    players_to_vehicle = models.PlayersToVehicleInExecution.query.filter_by(
         execution_id=execution.id,
         vehicle_name=vehicle,
-        player_tan=f"empty-{vehicle}"
-    ).first()
-    if player_to_vehicle:
-        # assign empty seat in vehicle
+    ).all()
+    if len(players_to_vehicle) > 0:
+
+        player_to_vehicle = players_to_vehicle[0]
+
+        tan = str(unique())
         player = (models.Player(tan=tan, execution_id=execution.id, # type: ignore
                                 location_id=player_to_vehicle.location_id, # type: ignore
                                 role_id=role,  # type: ignore
                                 alerted=False))  # type: ignore
-        db.session.add(player)
-        player_to_vehicle.player_tan = tan
-        db.session.commit()
 
-        # create new empty seat in vehicle
-        # Here it is possible to implement a limitation of seats if required
         new_seat_in_vehicle = models.PlayersToVehicleInExecution(
-            execution_id=player_to_vehicle.execution_id,  # type: ignore
+            execution_id=execution.id,  # type: ignore
             scenario_id=player_to_vehicle.scenario_id,  # type: ignore
-            player_tan=f"empty-{player_to_vehicle.vehicle_name}", # type: ignore
+            player_tan=tan,  # type: ignore
             location_id=player_to_vehicle.location_id,  # type: ignore
             vehicle_name=player_to_vehicle.vehicle_name,  # type: ignore
             travel_time=player_to_vehicle.travel_time)  # type: ignore
-        db.session.add(new_seat_in_vehicle)
 
+        db.session.add(player)
+        db.session.add(new_seat_in_vehicle)
         db.session.commit()
 
         # load player location -> choose from existing or load new vehicle
@@ -297,9 +342,7 @@ def __add_new_player_to_execution(execution: Execution, role: int,
             player_to_vehicle.vehicle_name)
         if not location:
             location = load_location(player_to_vehicle.location_id)
-        new_player = Player(tan, None, False, 0,
-                            location,
-                            set(), load_role(role))
+        new_player = Player(tan, None, False, 0, location, set(), load_role(role))
 
         # Add player to execution
         execution.players[tan] = new_player
@@ -307,7 +350,8 @@ def __add_new_player_to_execution(execution: Execution, role: int,
         if execution.status is execution.Status.PENDING or execution.Status.RUNNING:
             register_player(execution.id, [new_player])
     else:
-        msg = f"Unable to assign player to vehicle. No entry found for:{execution.id}{vehicle}"
+        msg = (f"Unable to assign player to vehicle. No vehicle found for:"
+               f"{execution.id}{vehicle}")
         logging.error(msg)
         raise BadRequest(msg)
